requirements.txt
torch==2.11.0
torchvision==0.26.0
timm==1.0.26
numpy==2.4.2
Pillow==12.2.0
matplotlib==3.10.8
tqdm==4.67.3
